+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Updated calls to `/mute` endpoint so they use request body instead of query params [#6702](https://github.com/ethyca/fides/pull/6702)
 - Changed default sort order for discovered assets from compliance status to asset name [#6704](https://github.com/ethyca/fides/pull/6704)
 - Break system table actions out of overflow menu [#6696](https://github.com/ethyca/fides/pull/6696)
+- Tweaks to system table UX [#6712](https://github.com/ethyca/fides/pull/6712)
 
 ### Developer Experience
 - Improved accessibility configuration for ESLint jsx-a11y plugin with Ant Design component support [#6699](https://github.com/ethyca/fides/pull/6699)